requirements.txt
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
oauthlib==3.2.2
py-trello==0.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.4
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
urllib3==2.0.4
